2-3: except Exception as var
python2 -3 warns about python2 construct which are deprecated for
python2 because they are incompatibly with python3. Replace them to get
as close as possible to a python2+python3 codebase.

First step:
Turn "except Exception, var" into "except Exception as var", which
removes the first error when running impressive with python3.



git-svn-id: http://svn.emphy.de/impressive/branches/python3/impressive@249 58003caa-664d-4c56-a378-6d187ff66ee1