2-3: generators and lists
filter and map return lists in py2, generators in py3. Both are
iterables. Check where we specifically need a list or tuple (and make
that explicit), or where an iterable is fine.

Note:
filter(None, iterable) is not a noop (it filters out "false"/"") so we
cannot replace it by iterable, unless we feed it e.g. into join.



git-svn-id: http://svn.emphy.de/impressive/branches/python3/impressive@259 58003caa-664d-4c56-a378-6d187ff66ee1