2-3: list/iterator idioms
Use values() for itervalues() etc. so that the code runs in py2 and py3.
In effect, this makes the code run on py3 the way it used to run on py2
(using an iterator) while it possible degrades performance on py2
(switching to a list). There seems to be no severe penalty, though.



git-svn-id: http://svn.emphy.de/impressive/branches/python3/impressive@264 58003caa-664d-4c56-a378-6d187ff66ee1