more crash fixes
- sort() with comparison function (instead of key)
- bytes vs. str in persistent cache

Note: persistent cache works now, but the cache files
themselves are *not* exchangeable between Python 2 and 3,
i.e. Impressive on Python 2 will not recognize a cache
file created with Python 3 as valid and vice-versa.
This is due to a different precision of the stat()
timestamps and their string conversions, which are
then hashed to get the cache key.


git-svn-id: http://svn.emphy.de/impressive/branches/python3/impressive@271 58003caa-664d-4c56-a378-6d187ff66ee1