workaround against crashes from loading the wrong SDL lib
In some cases, when installing PyGame through pip,
it ships with its own copy of libSDL. In this case,
loading the system's default SDL (to look up
OpenGL functions with SDL_GL_GetProcAddress) is fatal,
because this creates a *new* SDL instance unrelated
to PyGame's. We need to detect this and load PyGame's
own library instead.


git-svn-id: http://svn.emphy.de/impressive/branches/python3/impressive@284 58003caa-664d-4c56-a378-6d187ff66ee1